requirements.txt
numpy~=2.1.2
pandas~=2.2.3
requests~=2.32.3
python-dotenv~=1.0.1
Faker~=35.2.0
emoji~=2.14.1
scikit-learn~=1.6.1
transformers~=4.50.0
langchain_gigachat
openai
